Switched to using stable steam api where it's possible
